+++ docs/CHANGELOG.md
@@ -1,3 +1,4 @@
+- #1299 Updates tests in attester-incentives
 - #1275 Adds an `ExecutionContext` enum to `stf::apply_slot` and `Context::new`. This enum allows callees to distinguish between sequencer execution and normal "full node" execution.
 - #1290 Refactors slashing mechanism in the prover incentives module.
 - #1287 Breaking Change: Adds `FromStr` trait bound to a `BlockHashTrait`.